frontend/requirements.txt
fastapi==0.115.11
matplotlib==3.10.1
numpy==1.26.2
pandas==2.2.3
python-dotenv==1.0.1
pyvis==0.3.2
Requests==2.32.3
shutup==0.2.0
sseclient-py==1.7.2
streamlit==1.42.2
uvicorn==0.29.0
